requirements.txt
pydantic>=2.8,<3
python-dateutil>=2.9,<3
pandas>=2.2,<3
pyyaml>=6,<7
matplotlib>=3.9,<4
openpyxl>=3.1,<4
# Agent framework (optional for demo); installed but we fallback if API not configured
pyautogen
# OpenAI SDK (optional)
openai>=1.40,<2
streamlit>=1.38,<2
fastapi>=0.110,<1
uvicorn>=0.30,<1
numpy>=1.26,<3
scikit-learn>=1.4,<2
